New platform  for chip tii3qw81
